fix(build) : check duplicated job on compile()
1. Problem

Compile of latexmk is a bit time-consuming process.
If user executes compile again while a file is compiled, the compile
will be failed.

2. Solution

Check job_id which is created when jobstart() executes, it will be reset
to nil when the job is completed.
Compilation which is started before completion is ignored.