fix(utils) : replace sep_change() to sep_unify() and fix feature
Previously, the method of changing the path separator was determined by the OS.
Sometimes It is needed to change path separator to what we want, not by os.
Because paths of tex file will be used for key of server mapping file.

So the function name is changed.
sep_unify(path) -> change separator by OS
sep_unify(path, '/') -> change separator by '/'