feat : Add error pattern that starts with '!'
1. Problem - some ! error cannot be converted the format while
   `-file-line-error` is set

Previous method to detect error is catching pattern in the format of
'<filename>:<line>:<msg>' from log file. It needs to set
`-file-line-error` option for `latexmk`.

I found that some error message such as "<package> not found" doesn't
be converted its format to `<filename>:<line>:<msg>` even though
`-file-line-error` was set.
It makes texflow miss an error detection so It is the reason why I add
pattern that starts with '!', original error pattern.

But the original error pattern doesn't have line number, the line number
in the format of 'l.<line> <msg>' is related with error message above
it. So we needs to connect their relation while error is detecting.

2. Problem - Slicing chunk can break the relation between error and line
   number

Slicing chunk for multiprocessing needs to split exactly what I want.
Because we will use '!' pattern to check error, we needs to check the
next 'l.<line>' pattern. If the chunk split them, the relation will be
broken.

It may be implemented in splitting chunk function, but It is more
complex than previous method. so just use one core to parser.

If I feel it is too slow, I'll find the solution to proper splitting chunk.

3. Solution

Check '!' pattern and insert line number inside of message from `l.<line>` pattern.

1) add error pattern which start with '!'
2) use one chunk to parse log file
3) remove `filename` field in diagnostics because it isn't correct field