fix : separate `cleanup_auxfiles()` from `compile()`
1) Problem

When recompiling in onSave mode compilation, I was unsure whether to implement
it so that it deletes and recompiles every time or just do it as is.
It will makes compile time longer unnecessarily.

`latexmk` has feature that re-compile If there are incomplete compilation results.
So It may be not used often to remove aux files.

2) Solution

    1) Separate `cleanup_auxfiles()` function. Compile() always do compile.

    2) `Utils.scan_dir()` is created to find files matched with required
       patterns in base directory.
       Previous `vim.fn.glob()` seems to be slow than `vim.uv.fs_scandir`.
       It is applied to `search_documentclass()` also.
       To do this, `clear_ext` option accepts regex format. It will
       compare with filename using regex, not extension comparison.

    3) deleting files could be slow process. so It is implemented by
       asynchronous method using vim.uv.fs_unlink.
       If the deleting operation is ended, complete message will be
       shown.