fix(build) : fix wrong failed message when tex and pdf file is closed
1) Problem

When user close a tex file which is related with pdf file,
the pdf file is closed also.
Then `Failed to open viewer` error message invokes because the viewer job is terminated
and executes `on_exit` code.

2) Solution

The failed message is shown only when the `job_id.viewer` is valid.