feat(build) : onSave mode can recompile for changgin all tex/bib file in project
1) Problem

I set the autocmd to recompile for current buffer.
To simulate `-pvc` mode of `latexmk`, It needs to extend range to check
all files change in project.

2) solution

Most of user will edit *.tex or *.bib to work with latex.
`scan_dir()` is created to find required these patterns in root directory.
If some files in root directory is changed in neovim, rerun compile.

It has limitations unlike -pvc mode, The change must be made within
neovim.