fix(diagnostic) : show relative of maindir when warn_pkg if the file is in maindir
1) Problem

'warn_pkg' occurs some packages has warning. Most of packages are declared
in main tex file. but the warning is belonging to subfile or .sty file.

If .sty file, the file path will be shown properly with absolute path
format.
If it is subfile, diagnostics are shown in subfile, not main file which
is declared in. It prevent that user catches the diagnostics

2) Solution

When pkg_warn occurs, check not the file path is main file. it will
catch all files which is subfile or .sty file.
Then, If the file is in project root, use relative path, If not, show
relative path .