feat(diagnostic) : add `warn_nofile` case
New diagnostic case!
If reference filename is wrong, such as you add
`\include{chapter1.tex}`, It make .bbl file from the filename.
It will find chapter1.tex.bbl and it cannot be found because the
extension is wrong.

you should input the filename as \include{chapter1}

In this situation, `No file ~` warning is invoked. I didn't recognize
about it.